+++ src/app/mypage/editUser/page.tsx
@@ -1,5 +1,5 @@
-import UserEdit from '@/components/mypage/UserEdit';
+import UserEditContainer from '@/containers/mypage/UserEditContainer';
 
 export default function EditUser() {
-  return <UserEdit />;
+  return <UserEditContainer />;
 }
